Refactor: separate the charge.cpp and charge_mixing_test.cpp
